+++ packages/terra-form-select/tests/wdio/_frame-spec.js
@@ -1,7 +1,7 @@
 Terra.describeViewports('Frame', ['tiny'], () => {
-  describe('should have a themeable disabled icon', () => {
-    before(() => browser.url('/#/raw/tests/terra-form-select/form-select/disabled'));
+  it('should have a themeable disabled icon', () => {
+    browser.url('/raw/tests/terra-form-select/form-select/disabled');
 
-    Terra.it.validatesElement();
+    Terra.validates.element('themeable disabled icon');
   });
 });